[Misc] Move DP for ViT code inside model executor dir
